Clarify code & make button processing efficient
Instead of running the same function on several parameters (i.e. checking each
digit 1-9 separately and then checking each operand), this patch will
efficiently check the type of event and then respond to it by updating the
appropriate operands or operator list.

Some helper functions are removed as a result.

In addition, the 'ANS' function has been implemented: the accumulator (result of
an arithmetic expression) is displayed in its own text box as part of the UI.

Finally, '0' has been added as a valid operand, which hadn't been supported before.